feat: add selectionType prop to Calendar component
